--- src/main/java/produce365/care/DataSource.java
@@ -1,27 +1,0 @@
-package produce365.care;
-
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
-
-public class DataSource {
-	private static final String DRIVER = "oracle.jdbc.driver.OracleDriver";
-	private static final String URL = "jdbc:oracle:thin:@localhost:1521:xe";
-	private static final String USERNAME = "produce365";
-	private static final String PASSWORD = "a1234";
-	
-	public static Connection getDataSource() {
-		
-		Connection connection = null;
-		
-		try {
-		Class.forName(DRIVER);
-		connection = DriverManager.getConnection(URL, USERNAME, PASSWORD);
-		}catch(SQLException e) {
-			e.printStackTrace();
-		}catch(ClassNotFoundException e) {
-			e.printStackTrace();
-		}	
-		return connection;
-		}
-}
