+++ src/main/java/produce365/care/JDBCCareDAO.java
@@ -7,6 +7,8 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
+import produce365.common.DataSource;
+
 public class JDBCCareDAO implements CareDAO{
 
 	@Override
