+++ src/main/java/produce365/lesson/JDBCLessonDAO.java
@@ -7,7 +7,7 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-import produce365.care.DataSource;
+import produce365.common.DataSource;
 
 public class JDBCLessonDAO implements LessonDAO {
