+++ src/main/java/produce365/trainee/JDBCTraineeDAO.java
@@ -7,6 +7,8 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
+import produce365.common.DataSource;
+
 public class JDBCTraineeDAO implements TraineeDAO {
 
 	@Override
